Reworked lineage graph updates to use subscriptions service from the tab component instead of dataset component
