+++ src/app/components/dynamic-table/dynamic-table.interface.ts
@@ -1,4 +1,4 @@
-import { DataRow } from "src/app/dataset-view/datasetSubs.interface";
+import { DataRow } from "src/app/dataset-view/dataset.subscriptions.interface";
 import { DataSchemaField } from "../../interface/search.interface";
 
 export type TableSourceRowInterface = DataSchemaField | DataRow;
